+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `TestCaseUtils.SplitTestName` to split and ignore `(` and `)`.
 ### Tests
 - Add `TestsFixture` and `TestFixtureSource`.
+- Add `TestsRevitFixtureSource` with Revit parameter and `TestFixtureSource` attribute.
 
 ## [1.8.0] / 2024-12-16 - 2024-12-19
 ### Features